Hash improvements
Hash now has more functions to allow it to calculate hash codes for iterables. We take the size of collections into account as well when calculating hash codes for primitive arrays. The multipliers used are now static fields of Hash

Changed hashCode on ImLists to work on the iterator and consider the first 10 elements rather than converting it to an array of size 10.

ImMap - I now use the value in each entry to contribute to the hashcode of the Map. It is now cached. Changed hasEqualEntries() to iterate over both maps - should be much faster. I no longer return Entry objects from ImMap. They are converted to ImPairs.

ImSet - fixed hashCode() to prevent collisions more better-er. Added containsElementWhere() and findElementWhere()

ImValuesImpl - removed the obsolete hashed field and renamed the hash methods to match other classes